orvim-models/ocr/requirements.txt
fastapi
uvicorn
transformers
pillow
torch
protobuf
sentencepiece